refactor: move bulk of code into lib.rs for testing
