+++ Cargo.toml
@@ -7,6 +7,14 @@ readme = "README.md"
 repository = "https://github.com/Frazzer951/fpm"
 version = "0.4.1"
 
+[lib]
+name = "fpm"
+path = "src/lib.rs"
+
+[[bin]]
+name = "fpm"
+path = "src/main.rs"
+
 [dependencies]
 clap = { version = "3.2.6", features = ["cargo"] }
 config = "0.13.1"
@@ -23,3 +31,6 @@ serde_yaml = "0.9.0"
 strsim = "0.10.0"
 toml = "0.5.9"
 walkdir = "2.3.2"
+
+[dev-dependencies]
+serial_test = "0.8.0"
